[refactor]: PurchaseInfo 삭제
