+++ src/main/java/lotto/dto/input/PurchaseDto.java
@@ -4,21 +4,15 @@ import java.util.List;
 
 public class PurchaseDto {
     private final int purchaseMoney;
-    private final int numOfManualLottos;
     private final List<String> manualLottos;
 
-    public PurchaseDto(int purchaseMoney, int numOfManualLottos, List<String> manualLottos) {
+    public PurchaseDto(int purchaseMoney, List<String> manualLottos) {
         this.purchaseMoney = purchaseMoney;
-        this.numOfManualLottos = numOfManualLottos;
         this.manualLottos = manualLottos;
     }
 
     public int getPurchaseMoney() {
         return purchaseMoney;
-    }
-
-    public int getNumOfManualLottos() {
-        return numOfManualLottos;
     }
 
     public List<String> getManualLottos() {
