+++ src/main/java/lotto/view/InputView.java
@@ -22,7 +22,7 @@ public class InputView {
         int purchaseMoney = askPurchaseMoney();
         int numOfManualLottos = askNumOfManualLottos();
         List<String> manualLottos = askManualLotto(numOfManualLottos);
-        return new PurchaseDto(purchaseMoney, numOfManualLottos, manualLottos);
+        return new PurchaseDto(purchaseMoney, manualLottos);
     }
 
     public static WinningConditionDto getWinningConditionDto() {
